refactor: 관리자, 사용자 index 뷰 정리
